Refactor fd read/write code and add pread/pwrite

+++ implementations/meth2/libutil/file.hh
@@ -23,7 +23,7 @@ public:
   void fsync( void );
 
   /* file size */
-  size_t size( void ) const;
+  uint64_t size( void ) const;
 };
 
 #endif /* FILE_HH */
